More Axes-to-AxesBase refactor
 * Moved nearly all axis-agnostic methods to AxesBase
 * Added compatibility shims for self.xaxis and self.yaxis
 * Fixed PEP8 issues
 * Fixed outdated python-isms such as '<>'.
 * All hasattr(ax, 'xaxis') calls must now be replaced with
   ax.xaxis is not None and the same for yaxis.
